Types/Methods/Functions

Index

NmodController.CalciumDynamic

NmodController.IonCurrent

NmodController.NeuronCB

NmodController.computeDICs

NmodController.computeThresholdVoltage

NmodController.initializeCalciumDynamics

NmodController.initializeCurrent

NmodController.initializeNeuronModel

NmodController.writeControlledODEs

NmodController.writeUncontrolledODEs

Public Interface

NmodController.CalciumDynamic — Type

struct CalciumDynamic
    numberOfCurrents::Int64
    currentNames::Vector{String}
    coefficients::Vector{Float64}
    nernstEquilibriumValue::Float64
    timeConstant::Float64
end

Data structure defining an ODE for intracellular calcium dynamics.

To initialize a certain calcium dynamics, please use the function initializeCalciumDynamics.

NmodController.IonCurrent — Type

struct IonCurrent
    name::String
    numberOfGatings::Int64
    exponents::Vector{Int64}
    steadyStateGatings::Vector{Function}
    timeConstants::Vector{Function}
    reversalPotential::Float64
    calciumDependency::Bool
    MgDependency::Bool
end

Data structure defining an ionic current in a conductance based model.

To initialize a certain type of current, please use the function initializeCurrent.

NmodController.NeuronCB — Type

struct NeuronCB
    C::Float64
    numberOfIonCurrents::Int64
    ionCurrents::Vector{IonCurrent}
    maximumConductances::Vector{Float64}
    leakageConductance::Float64
    reversaleLeakagePotential::Float64
    calciumDynamics::Union{CalciumDynamic, Bool}
    globalCalciumDependency::Bool
    globalMgDependency::Bool
end

Data structure defining a complete conductance based model.

To initialize a certain type of model, please use the function initializeNeuronModel.

NmodController.computeDICs — Method

computeDICs(neuron::NeuronCB, 
    tauFast::Function, 
    tauSlow::Function, 
    tauUltraslow::Function; 
    tauCa::Union{Float64, Int64, Function}=NaN, 
    Mg::Union{Int64, Float64}=NaN, 
    onlyS::Bool=false, 
    scaled::Bool=true)

Compute dynamic input conductances gDIC (DICs) or only the sensitivity matrix S of a conductance based model. Note that gDIC = S*barg with barg being the vector of all maximum ion channel conductances. Refer to Drion et al., 2015 "Dynamic Input Conductances Shape Neuronal Spiking"

Arguments

neuron: data structure containing the conductance based model of interest.

tauFast: fast time constant function, used to compute the DIC weights.

tauSlow: slow time constant function, used to compute the DIC weights.

tauUltraslow: ultraslow time constant function, used to compute the DIC weights.

tauCa: time constant of the calcium if the model is calcium dependent, used to the DIC weights for the calcium only. Optional. Default value is NaN. Required if the neuron model is calcium dependent.

Mg: magnesium concentration if the model is magnesium dependent. Optional. Default value is NaN. Required if the neuron model is magnesium dependent.

onlyS: flag indicating if only the sensitivity matrix should be computed, for the case where no maximum ion channel conductances are in the model data structure. Optional. Default value is false.

scaled: flag indicating if DICs or sensitivity matrix should be scaled by the leakage conductance. Optional. Default value is true.

Example

julia> gf, gs, gu = computeDICs(STG, STG_tauFast, STG_tauSlow, STG_tauUltraslow, tauCa=5000.);

julia> [gf(0), gs(0), gu(0)]
3-element Vector{Float64}:
  2712.186654147375
 19069.69205128652 
  -196.85745712674108

julia> S = computeDICs(STG, STG_tauFast, STG_tauSlow, STG_tauUltraslow, tauCa=5000., onlyS=true);

julia> S(0)
3×7 Matrix{Float64}:
 0.0213593  0.23756  0.00590276   0.000796009   0.0766643   29.8745   0.000857487
 0.059507   2.47411  0.0353302   -0.00824027    0.118022   211.202   -0.0
 0.0        1.29588  0.0386544   -0.00383275   -3.34257      0.0     -0.00285826

NmodController.computeThresholdVoltage — Method

computeThresholdVoltage(gf::Function, 
    gs::Function, 
    gu::Function)

Compute threshold voltage of a specific conductance based model. Refer to Drion et al., 2015 "Dynamic Input Conductances Shape Neuronal Spiking"

Arguments

gf: fast DIC. Must be of type Function.

gs: slow DIC. Must be of type Function.

gu: ultraslow DIC. Must be of type Function.

Example

julia> Vth = computeThresholdVoltage(gf, gs, gu)
-48.897088456154556

NmodController.initializeCalciumDynamics — Method

initializeCalciumDynamics(currentNames::Union{String, Vector{String}}, 
    coefficients::Union{Int64, Float64, Vector{Int64}, Vector{Float64}},
    nernstEquilibriumValue::Union{Int64, Float64}, 
    timeConstant::Union{Int64, Float64})

Initialize an intracellular calcium dynamics data structure. The ODE describing intracellular calcium dynamics must be of type: tauCa * dCa/dt = (e1 * Iion1 + ... + en * Iion_n - Ca + CaEquilibrium).

Arguments

currentNames: names of the calcium ionic current Iion_i. Must be of the same length as coefficients.

coefficients: coefficient(s) of the ionic currents, ei. Must be of the same length as currentNames.

nernstEquilibriumValue: Equilibrium value of the intracellular calcium with null current, CaEquilibrium.

timeConstant: time constant of the intracellular calcium dynamics, tau_Ca.

Example

julia> CaDynamics = initializeCalciumDynamics(["CaT", "CaS"], [-0.94, -0.94], 0.05, 20)

CalciumDynamic(2, ["CaT", "CaS"], [-0.94, -0.94], 0.05, 20.0)

NmodController.initializeCurrent — Method

initializeCurrent(name::String, 
    reversalPotential::Union{Int64, Float64}; 
    numberOfGatings::Int64=1, 
    exponents::Union{Int64, Vector{Int64}}=1, 
    activationSteadyStateGating::Union{Function, Float64}=0., 
    activationTimeConstant::Union{Function, Float64, Int64}=NaN, 
    inactivationSteadyStateGating::Union{Function, Float64}=0., 
    inactivationTimeConstant::Union{Function, Float64, Int64}=NaN, 
    calciumDependency::Bool=false, 
    MgDependency::Bool=false)

Initialize an ionic current data structure. The current must be of type: Iion = bargion * m^a1 * h^a2 * (V - Eion). The ODE describing gating variable dynamics must be of type: taum(V) * dm/dt = m_inf(V, [Ca, Mg]) - m.

Note that the maximum ion channel conductance bar_gion is not contained in the ionic current data structure but in the neuronal model one.

Arguments

name: name of the ionic current.

reversalPotential: reversal Nernst potential of the ion, Eion.

numberOfGatings: number of gating variable(s) of the ionic current. Optional. Default value is 1.

exponents: exponent(s) of the gating variable(s), a1 (and a2). Optional. Default value is 1. Must be a Vector of length 2 if numberOfGatings = 2.

activationSteadyStateGating: equilibrium function of the activation gating variable. Optional. Default value is 0. Must be of type Function, always required.

activationTimeConstant: time constant function/constant of the activation gating variable. Optional. Default value is NaN. Must be of type Function or an Int64/Float64 different from NaN, always required.

inactivationSteadyStateGating: equilibrium function of the inactivation gating variable. Optional. Default value is 0. Must be of type Function, required if numberOfGatings = 2.

inactivationTimeConstant: time constant function/constant of the inactivation gating variable. Optional. Default value is NaN. Must be of type Function or Int64/Float64 different from NaN, required if numberOfGatings = 2.

calciumDependency: flag indicating if the ionic current depends on calcium. Optional. Default value is false. Required if activationSteadyStateGating second argument is calcium.

MgDependency: flag indicating if the ionic current depends on magnesium. Optional. Default value is false. Required if activationSteadyStateGating second argument is magnesium.

Example

julia> NaCurrent = initializeCurrent("Na", VNa, numberOfGatings=2, exponents=[3, 1],
    activationSteadyStateGating=mNa_inf, activationTimeConstant=tau_mNa,
    inactivationSteadyStateGating=hNa_inf, inactivationTimeConstant=tau_hNa)
    
IonCurrent("Na", 2, [3, 1], Function[mNa_inf, hNa_inf], Function[tau_mNa, tau_hNa], 50.0, false, false)

NmodController.initializeNeuronModel — Method

initializeNeuronModel(ionCurrents::Union{IonCurrent, Vector{IonCurrent}}; 
    C::Float64=1., 
    leakageConductance::Union{Float64, Int64}=1., 
    reversaleLeakagePotential::Union{Float64, Int64}=-50., 
    calciumDynamics::Union{CalciumDynamic, Bool}=false, 
    maximumConductances::Union{Vector{Int64}, Int64, Vector{Float64}, Float64}=0.)

Initialize a conductance based model data structure. The leakage current must be of type: Ileak = gleak * (V - Eleak).

Arguments

ionCurrents: vector of previously defined active ion current data structures contained in the model.

C: membrane capacitance. Optional. Default values is 1.

leakageConductance: conductance of the leakage current, gleak. Optional. Default value is 1.

reversaleLeakagePotential: reversal Nernst potential of the leakage current, Eleak. Optional. Default value is -50.

calciumDynamics: data structure defining intracellular calcium dynamics. Optional. Must be of type CalciumDynamic is at least one ionic current is calcium dependent.

maximumConductances: vector of maximum ion channel conductances of currents contained in ionCurrents. Optional. Default value is a vector full of NaNs. Must be of the same length as ionCurrents.

Example

julia> STG = initializeNeuronModel(STG_ionCurrents, C=0.1, calciumDynamics=CaDyn, 
    leakageConductance=0.01, reversaleLeakagePotential=STG_Vleak, maximumConductances=STG_gvec)

NeuronCB(0.1, 7, IonCurrent[IonCurrent("Na", 2, [3, 1], Function[STG_mNa_inf, STG_hNa_inf], Function[STG_tau_mNa, STG_tau_hNa], 50.0, false, false), 
IonCurrent("CaT", 2, [3, 1], Function[STG_mCaT_inf, STG_hCaT_inf], Function[STG_tau_mCaT, STG_tau_hCaT], 80.0, false, false), 
IonCurrent("CaS", 2, [3, 1], Function[STG_mCaS_inf, STG_hCaS_inf], Function[STG_tau_mCaS, STG_tau_hCaS], 80.0, false, false), 
IonCurrent("A", 2, [3, 1], Function[STG_mA_inf, STG_hA_inf], Function[STG_tau_mA, STG_tau_hA], -80.0, false, false), 
IonCurrent("KCa", 1, [4], Function[STG_mKCa_inf], Function[STG_tau_mKCa], -80.0, true, false), 
IonCurrent("Kd", 1, [4], Function[STG_mKd_inf], Function[STG_tau_mKd], -80.0, false, false), 
IonCurrent("H", 1, [1], Function[STG_mH_inf], Function[STG_tau_mH], -20.0, false, false)], 
[800.0, 3.0, 3.0, 80.0, 60.0, 90.0, 0.1], 0.01, -50.0, CalciumDynamic(2, ["CaT", "CaS"], [-0.94, -0.94], 0.05, 20.0), true, false)

NmodController.writeControlledODEs — Method

writeControlledODEs(neuron::NeuronCB, 
    controlledConductances::Vector{String}, 
    controlledDICs::Vector{String}; 
    filename::String="ControlledCBModelODEs.jl")

Write a julia file that contains function computing all ODEs for a specific conductance based model that is coupled with a neuromodulation controller. The function have the same name as the file, without the ".jl" extension. The equations can be integrated using the DifferentialEquations.jl package. Refer to Fyon et al., 2023 "Reliable neuromodulation from adaptive control of ion channel expression". Returns nothing.

Arguments

neuron: data structure containing the conductance based model of interest.

controlledConductances: vector of strings that contains the names of the conductances the controller outputs. Its length must be higher or equal then the length of controlledDICs.

controlledDICs: vector of strings that contains the names of the DICs the controller gets in input. Its length must be lower or equal than the length of controlledConductances.

filename: name of the julia file that will contain the equations. Optional. Default value is "ControlledCBModelODEs.jl".

Example

julia> writeControlledODEs(STG, ["CaS", "A"], ["s", "u"])

NmodController.writeUncontrolledODEs — Method

writeUncontrolledODEs(neuron::NeuronCB; 
    filename::String="CBModelODEs.jl")

Write a julia file that contains a function computing all ODEs for a specific conductance based model. The function have the same name as the file, without the ".jl" extension. The equations can be integrated using the DifferentialEquations.jl package. Returns nothing.

Arguments

neuron: data structure containing the conductance based model of interest.

filename: name of the julia file that will contain the equations. Optional. Default value is "CBModelODEs.jl".

Example

julia> writeUncontrolledODEs(STG)